Module aws_sdk_backup::error

Modules

already_exists_exception

See AlreadyExistsException

conflict_exception

See ConflictException

dependency_failure_exception

See DependencyFailureException

invalid_parameter_value_exception

See InvalidParameterValueException

invalid_request_exception

See InvalidRequestException

invalid_resource_state_exception

See InvalidResourceStateException

limit_exceeded_exception

See LimitExceededException

missing_parameter_value_exception

See MissingParameterValueException

resource_not_found_exception

See ResourceNotFoundException

service_unavailable_exception

See ServiceUnavailableException

Structs

AlreadyExistsException

The required resource already exists.

ConflictException

Backup can't perform the action that you requested until it finishes

CreateBackupPlanError

CreateBackupSelectionError

CreateBackupVaultError

CreateFrameworkError

CreateReportPlanError

DeleteBackupPlanError

DeleteBackupSelectionError

DeleteBackupVaultAccessPolicyError

DeleteBackupVaultError

DeleteBackupVaultNotificationsError

DeleteFrameworkError

DeleteRecoveryPointError

DeleteReportPlanError

DependencyFailureException

A dependent Amazon Web Services service or resource returned an error to the Backup service, and the action cannot be completed.

DescribeBackupJobError

DescribeBackupVaultError

DescribeCopyJobError

DescribeFrameworkError

DescribeGlobalSettingsError

DescribeProtectedResourceError

DescribeRecoveryPointError

DescribeRegionSettingsError

DescribeReportJobError

DescribeReportPlanError

DescribeRestoreJobError

DisassociateRecoveryPointError

ExportBackupPlanTemplateError

GetBackupPlanError

GetBackupPlanFromJSONError

GetBackupPlanFromTemplateError

GetBackupSelectionError

GetBackupVaultAccessPolicyError

GetBackupVaultNotificationsError

GetRecoveryPointRestoreMetadataError

GetSupportedResourceTypesError

InvalidParameterValueException

Indicates that something is wrong with a parameter's value. For example, the value is

InvalidRequestException

Indicates that something is wrong with the input to the request. For example, a

InvalidResourceStateException

Backup is already performing an action on this recovery point. It can't

LimitExceededException

A limit in the request has been exceeded; for example, a maximum number of items allowed

ListBackupJobsError

ListBackupPlanTemplatesError

ListBackupPlanVersionsError

ListBackupPlansError

ListBackupSelectionsError

ListBackupVaultsError

ListCopyJobsError

ListFrameworksError

ListProtectedResourcesError

ListRecoveryPointsByBackupVaultError

ListRecoveryPointsByResourceError

ListReportJobsError

ListReportPlansError

ListRestoreJobsError

ListTagsError

MissingParameterValueException

Indicates that a required parameter is missing.

PutBackupVaultAccessPolicyError

PutBackupVaultNotificationsError

ResourceNotFoundException

A resource that is required for the action doesn't exist.

ServiceUnavailableException

The request failed due to a temporary failure of the server.

StartBackupJobError

StartCopyJobError

StartReportJobError

StartRestoreJobError

StopBackupJobError

TagResourceError

UntagResourceError

UpdateBackupPlanError

UpdateFrameworkError

UpdateGlobalSettingsError

UpdateRecoveryPointLifecycleError

UpdateRegionSettingsError

UpdateReportPlanError

Enums

CreateBackupPlanErrorKind

CreateBackupSelectionErrorKind

CreateBackupVaultErrorKind

CreateFrameworkErrorKind

CreateReportPlanErrorKind

DeleteBackupPlanErrorKind

DeleteBackupSelectionErrorKind

DeleteBackupVaultAccessPolicyErrorKind

DeleteBackupVaultErrorKind

DeleteBackupVaultNotificationsErrorKind

DeleteFrameworkErrorKind

DeleteRecoveryPointErrorKind

DeleteReportPlanErrorKind

DescribeBackupJobErrorKind

DescribeBackupVaultErrorKind

DescribeCopyJobErrorKind

DescribeFrameworkErrorKind

DescribeGlobalSettingsErrorKind

DescribeProtectedResourceErrorKind

DescribeRecoveryPointErrorKind

DescribeRegionSettingsErrorKind

DescribeReportJobErrorKind

DescribeReportPlanErrorKind

DescribeRestoreJobErrorKind

DisassociateRecoveryPointErrorKind

ExportBackupPlanTemplateErrorKind

GetBackupPlanErrorKind

GetBackupPlanFromJSONErrorKind

GetBackupPlanFromTemplateErrorKind

GetBackupSelectionErrorKind

GetBackupVaultAccessPolicyErrorKind

GetBackupVaultNotificationsErrorKind

GetRecoveryPointRestoreMetadataErrorKind

GetSupportedResourceTypesErrorKind

ListBackupJobsErrorKind

ListBackupPlanTemplatesErrorKind

ListBackupPlanVersionsErrorKind

ListBackupPlansErrorKind

ListBackupSelectionsErrorKind

ListBackupVaultsErrorKind

ListCopyJobsErrorKind

ListFrameworksErrorKind

ListProtectedResourcesErrorKind

ListRecoveryPointsByBackupVaultErrorKind

ListRecoveryPointsByResourceErrorKind

ListReportJobsErrorKind

ListReportPlansErrorKind

ListRestoreJobsErrorKind

ListTagsErrorKind

PutBackupVaultAccessPolicyErrorKind

PutBackupVaultNotificationsErrorKind

StartBackupJobErrorKind

StartCopyJobErrorKind

StartReportJobErrorKind

StartRestoreJobErrorKind

StopBackupJobErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateBackupPlanErrorKind

UpdateFrameworkErrorKind

UpdateGlobalSettingsErrorKind

UpdateRecoveryPointLifecycleErrorKind

UpdateRegionSettingsErrorKind

UpdateReportPlanErrorKind