Module aws_sdk_backup::model

Modules

advanced_backup_setting

See AdvancedBackupSetting

backup_job

See BackupJob

backup_plan

See BackupPlan

backup_plan_input

See BackupPlanInput

backup_plan_templates_list_member

See BackupPlanTemplatesListMember

backup_plans_list_member

See BackupPlansListMember

backup_rule

See BackupRule

backup_rule_input

See BackupRuleInput

backup_selection

See BackupSelection

backup_selections_list_member

See BackupSelectionsListMember

backup_vault_list_member

See BackupVaultListMember

calculated_lifecycle

See CalculatedLifecycle

condition

See Condition

control_input_parameter

See ControlInputParameter

control_scope

See ControlScope

copy_action

See CopyAction

copy_job

See CopyJob

framework

See Framework

framework_control

See FrameworkControl

lifecycle

See Lifecycle

protected_resource

See ProtectedResource

recovery_point_by_backup_vault

See RecoveryPointByBackupVault

recovery_point_by_resource

See RecoveryPointByResource

recovery_point_creator

See RecoveryPointCreator

report_delivery_channel

See ReportDeliveryChannel

report_destination

See ReportDestination

report_job

See ReportJob

report_plan

See ReportPlan

report_setting

See ReportSetting

restore_jobs_list_member

See RestoreJobsListMember

Structs

AdvancedBackupSetting

A list of backup options for each resource type.

BackupJob

Contains detailed information about a backup job.

BackupPlan

Contains an optional backup plan display name and an array of BackupRule

BackupPlanInput

Contains an optional backup plan display name and an array of BackupRule

BackupPlanTemplatesListMember

An object specifying metadata associated with a backup plan template.

BackupPlansListMember

Contains metadata about a backup plan.

BackupRule

Specifies a scheduled task used to back up a selection of resources.

BackupRuleInput

Specifies a scheduled task used to back up a selection of resources.

BackupSelection

Used to specify a set of resources to a backup plan.

BackupSelectionsListMember

Contains metadata about a BackupSelection object.

BackupVaultListMember

Contains metadata about a backup vault.

CalculatedLifecycle

Contains DeleteAt and MoveToColdStorageAt timestamps, which

Condition

Contains an array of triplets made up of a condition type (such as

ControlInputParameter

A list of parameters for a control. A control can have zero, one, or more than one

ControlScope

A framework consists of one or more controls. Each control has its own control scope.

CopyAction

The details of the copy operation.

CopyJob

Contains detailed information about a copy job.

Framework

Contains detailed information about a framework. Frameworks contain controls, which

FrameworkControl

Contains detailed information about all of the controls of a framework. Each framework

Lifecycle

Contains an array of Transition objects specifying how long in days before

ProtectedResource

A structure that contains information about a backed-up resource.

RecoveryPointByBackupVault

Contains detailed information about the recovery points stored in a backup vault.

RecoveryPointByResource

Contains detailed information about a saved recovery point.

RecoveryPointCreator

Contains information about the backup plan and rule that Backup used to

ReportDeliveryChannel

Contains information from your report plan about where to deliver your reports,

ReportDestination

Contains information from your report job about your report destination.

ReportJob

Contains detailed information about a report job. A report job compiles a report based

ReportPlan

Contains detailed information about a report plan.

ReportSetting

Contains detailed information about a report setting.

RestoreJobsListMember

Contains metadata about a restore job.

Enums

BackupJobState

BackupVaultEvent

ConditionType

CopyJobState

RecoveryPointStatus

RestoreJobStatus

StorageClass