Struct aws_sdk_backup::model::RecoveryPointByBackupVault

#[non_exhaustive]
pub struct RecoveryPointByBackupVault {

    pub recovery_point_arn: Option<String>,
    pub backup_vault_name: Option<String>,
    pub backup_vault_arn: Option<String>,
    pub source_backup_vault_arn: Option<String>,
    pub resource_arn: Option<String>,
    pub resource_type: Option<String>,
    pub created_by: Option<RecoveryPointCreator>,
    pub iam_role_arn: Option<String>,
    pub status: Option<RecoveryPointStatus>,
    pub status_message: Option<String>,
    pub creation_date: Option<Instant>,
    pub completion_date: Option<Instant>,
    pub backup_size_in_bytes: Option<i64>,
    pub calculated_lifecycle: Option<CalculatedLifecycle>,
    pub lifecycle: Option<Lifecycle>,
    pub encryption_key_arn: Option<String>,
    pub is_encrypted: bool,
    pub last_restore_time: Option<Instant>,

}

Contains detailed information about the recovery points stored in a backup vault.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

recovery_point_arn: Option<String>

An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

backup_vault_name: Option<String>

The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

backup_vault_arn: Option<String>

An ARN that uniquely identifies a backup vault; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

source_backup_vault_arn: Option<String>

The backup vault where the recovery point was originally copied from. If the recovery point is restored to the same account this value will be null.

resource_arn: Option<String>

An ARN that uniquely identifies a resource. The format of the ARN depends on the resource type.

resource_type: Option<String>

The type of Amazon Web Services resource saved as a recovery point; for example, an Amazon Elastic Block Store (Amazon EBS) volume or an Amazon Relational Database Service (Amazon RDS) database. For Windows Volume Shadow Copy Service (VSS) backups, the only supported resource type is Amazon EC2.

created_by: Option<RecoveryPointCreator>

Contains identifying information about the creation of a recovery point, including the BackupPlanArn, BackupPlanId, BackupPlanVersion, and BackupRuleId of the backup plan that is used to create it.

iam_role_arn: Option<String>

Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

status: Option<RecoveryPointStatus>

A status code specifying the state of the recovery point.

status_message: Option<String>

A message explaining the reason of the recovery point deletion failure.

creation_date: Option<Instant>

The date and time a recovery point is created, in Unix format and Coordinated Universal Time (UTC). The value of CreationDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

completion_date: Option<Instant>

The date and time a job to restore a recovery point is completed, in Unix format and Coordinated Universal Time (UTC). The value of CompletionDate is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

backup_size_in_bytes: Option<i64>

The size, in bytes, of a backup.

calculated_lifecycle: Option<CalculatedLifecycle>

A CalculatedLifecycle object containing DeleteAt and MoveToColdStorageAt timestamps.

lifecycle: Option<Lifecycle>

The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define.

Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Only Amazon EFS file system backups can be transitioned to cold storage.

encryption_key_arn: Option<String>

The server-side encryption key that is used to protect your backups; for example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

is_encrypted: bool

A Boolean value that is returned as TRUE if the specified recovery point is encrypted, or FALSE if the recovery point is not encrypted.

last_restore_time: Option<Instant>

The date and time a recovery point was last restored, in Unix format and Coordinated Universal Time (UTC). The value of LastRestoreTime is accurate to milliseconds. For example, the value 1516925490.087 represents Friday, January 26, 2018 12:11:30.087 AM.

Implementations

impl RecoveryPointByBackupVault

pub fn builder() -> Builder

Creates a new builder-style object to manufacture RecoveryPointByBackupVault