Module aws_sdk_backup::operation

Structs

CreateBackupPlan

Creates a backup plan using a backup plan name and backup rules. A backup plan is a

CreateBackupSelection

Creates a JSON document that specifies a set of resources to assign to a backup plan.

CreateBackupVault

Creates a logical container where backups are stored. A CreateBackupVault

CreateFramework

Creates a framework with one or more controls. A framework is a collection of controls

CreateReportPlan

Creates a report plan. A report plan is a document that contains information about the

DeleteBackupPlan

Deletes a backup plan. A backup plan can only be deleted after all associated selections

DeleteBackupSelection

Deletes the resource selection associated with a backup plan that is specified by the

DeleteBackupVault

Deletes the backup vault identified by its name. A vault can be deleted only if it is

DeleteBackupVaultAccessPolicy

Deletes the policy document that manages permissions on a backup vault.

DeleteBackupVaultNotifications

Deletes event notifications for the specified backup vault.

DeleteFramework

Deletes the framework specified by a framework name.

DeleteRecoveryPoint

Deletes the recovery point specified by a recovery point ID.

DeleteReportPlan

Deletes the report plan specified by a report plan name.

DescribeBackupJob

Returns backup job details for the specified BackupJobId.

DescribeBackupVault

Returns metadata about a backup vault specified by its name.

DescribeCopyJob

Returns metadata associated with creating a copy of a resource.

DescribeFramework

Returns the framework details for the specified FrameworkName.

DescribeGlobalSettings

Describes whether the Amazon Web Services account is opted in to cross-account backup.

DescribeProtectedResource

Returns information about a saved resource, including the last time it was backed up,

DescribeRecoveryPoint

Returns metadata associated with a recovery point, including ID, status, encryption, and

DescribeRegionSettings

Returns the current service opt-in settings for the Region. If service opt-in is enabled

DescribeReportJob

Returns the details associated with creating a report as specified by its

DescribeReportPlan

Returns a list of all report plans for an Amazon Web Services account and Amazon Web Services Region.

DescribeRestoreJob

Returns metadata associated with a restore job that is specified by a job ID.

DisassociateRecoveryPoint

Deletes the specified continuous backup recovery point from Backup and

ExportBackupPlanTemplate

Returns the backup plan that is specified by the plan ID as a backup template.

GetBackupPlan

Returns BackupPlan details for the specified BackupPlanId. The

GetBackupPlanFromJSON

Returns a valid JSON document specifying a backup plan or an error.

GetBackupPlanFromTemplate

Returns the template specified by its templateId as a backup plan.

GetBackupSelection

Returns selection metadata and a document in JSON format that specifies a list of

GetBackupVaultAccessPolicy

Returns the access policy document that is associated with the named backup

GetBackupVaultNotifications

Returns event notifications for the specified backup vault.

GetRecoveryPointRestoreMetadata

Returns a set of metadata key-value pairs that were used to create the backup.

GetSupportedResourceTypes

Returns the Amazon Web Services resource types supported by Backup.

ListBackupJobs

Returns a list of existing backup jobs for an authenticated account for the last 30

ListBackupPlanTemplates

Returns metadata of your saved backup plan templates, including the template ID, name,

ListBackupPlanVersions

Returns version metadata of your backup plans, including Amazon Resource Names (ARNs),

ListBackupPlans

Returns a list of all active backup plans for an authenticated account. The list

ListBackupSelections

Returns an array containing metadata of the resources associated with the target backup

ListBackupVaults

Returns a list of recovery point storage containers along with information about

ListCopyJobs

Returns metadata about your copy jobs.

ListFrameworks

Returns a list of all frameworks for an Amazon Web Services account and Amazon Web Services Region.

ListProtectedResources

Returns an array of resources successfully backed up by Backup, including

ListRecoveryPointsByBackupVault

Returns detailed information about the recovery points stored in a backup vault.

ListRecoveryPointsByResource

Returns detailed information about all the recovery points of the type specified by a

ListReportJobs

Returns details about your report jobs.

ListReportPlans

Returns a list of your report plans. For detailed information about a single report

ListRestoreJobs

Returns a list of jobs that Backup initiated to restore a saved resource,

ListTags

Returns a list of key-value pairs assigned to a target recovery point, backup plan, or

PutBackupVaultAccessPolicy

Sets a resource-based policy that is used to manage access permissions on the target

PutBackupVaultNotifications

Turns on notifications on a backup vault for the specified topic and events.

StartBackupJob

Starts an on-demand backup job for the specified resource.

StartCopyJob

Starts a job to create a one-time copy of the specified resource.

StartReportJob

Starts an on-demand report job for the specified report plan.

StartRestoreJob

Recovers the saved resource identified by an Amazon Resource Name (ARN).

StopBackupJob

Attempts to cancel a job to create a one-time backup of a resource.

TagResource

Assigns a set of key-value pairs to a recovery point, backup plan, or backup vault

UntagResource

Removes a set of key-value pairs from a recovery point, backup plan, or backup vault

UpdateBackupPlan

Updates an existing backup plan identified by its backupPlanId with the

UpdateFramework

Updates an existing framework identified by its FrameworkName with the

UpdateGlobalSettings

Updates whether the Amazon Web Services account is opted in to cross-account backup.

UpdateRecoveryPointLifecycle

Sets the transition lifecycle of a recovery point.

UpdateRegionSettings

Updates the current service opt-in settings for the Region. If service-opt-in is enabled

UpdateReportPlan

Updates an existing report plan identified by its ReportPlanName with the